zhou/pyproject.toml
[tool.poetry]
name = "zhou"
version = "0.1.0"
description = "Backend for Tonho School Management App"
authors = ["zimthug <tmlangeni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-restframework = "^0.0.1"
django = "^5.1.3"
psycopg2 = "^2.9.10"
python-decouple = "^3.8"
social-auth-app-django = "^5.4.2"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.3.1"
pillow = "^11.0.0"
django-cors-headers = "^4.6.0"
drf-spectacular = "^0.28.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
